refactor:  utils 파일 내 함수들을 분리해 동명의 파일 밑으로 옮겨 구조화
