+++ src/utils/gameLogic.js
@@ -1,23 +1,11 @@
+import { $ } from './dom.js';
 import { INPUT_RULE } from './constants.js';
-import Car from './Car.js';
-
-export const $ = (selector) => document.querySelector(selector);
+import Car from '../Car.js';
 
 export const alertError = (message, $element) => {
   $element.value = '';
   alert(message);
   $element.focus();
-};
-
-export const hideRacingCountForm = () => {
-  $('#racing-count-form-title').style.display = 'none';
-  $('#racing-count-form').style.display = 'none';
-};
-
-export const revealRacingCountForm = () => {
-  $('#racing-count-form-title').style.display = 'block';
-  $('#racing-count-form').style.display = 'block';
-  $('#racing-count-input').focus();
 };
 
 export const validateNames = (names) => {
@@ -53,15 +41,6 @@ export const createCars = (names) => {
   return cars;
 };
 
-export const createRacingResultArea = () => {
-  const AREA_ID = 'racing-result';
-
-  const $racingResultArea = document.createElement('p');
-  $racingResultArea.id = AREA_ID;
-
-  $('#app').appendChild($racingResultArea);
-};
-
 export const proceedRacingStage = (cars) => {
   cars.map((car) => {
     const randomNum = car.generateRandomNumber();
@@ -89,18 +68,4 @@ export const getWinners = (cars) => {
     ({ forwardCount }) => forwardCount === MAX_FORWARD
   );
   return winners;
-};
-
-export const paintWinners = (winners) => {
-  const WINNERS_NAMES = winners.map(({ name }) => name);
-
-  const $resultLine = document.createElement('span');
-  $resultLine.innerHTML = '최종 우승자: ';
-
-  const $racingWinners = document.createElement('span');
-  $racingWinners.id = 'racing-winners';
-  $racingWinners.innerHTML = WINNERS_NAMES.join();
-
-  $resultLine.appendChild($racingWinners);
-  $('#racing-result').appendChild($resultLine);
 };
